そっかぁ・・・intじゃないと負数のチェックは出来ないよねぇ・・・。
git-svn-id: http://svn.sourceforge.jp/svnroot/open-mgl/MGL/trunk@97 6d62ff88-fa28-0410-b5a4-834eb811a934

+++ mgllib/src/2d/MglBitmapData.cpp
@@ -123,7 +123,7 @@ void CMglBitmapData::Fill(D3DCOLOR color, RECT rect){
 		nEndY = m_nHeight;
 
 	//size_t nFillSize = (nStartX) * sizeof(D3DCOLOR);
-	size_t nFillSize = nEndX - nStartX + 1;
+	int nFillSize = nEndX - nStartX + 1;
 	if ( nFillSize <= 0 )
 		return;
